Refactoring answering model into a directory
