+++ qaeval/__init__.py
@@ -1,4 +1,4 @@
 from qaeval.answer_selection import AnswerSelector
-from qaeval.answering import QuestionAnsweringModel
+from qaeval.answering.model import QuestionAnsweringModel
 from qaeval.generation.model import QuestionGenerationModel
 from qaeval.version import VERSION as __version__
